requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2
psycopg2-binary==2.9.7
celery==5.3.4
redis==5.0.1
Pillow==10.1.0
django-filter==23.3
django-extensions==3.2.3
reportlab==4.0.7
openpyxl==3.1.2
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.6.0
